👙🐲 Updated with Glitch

+++ commands/poll.js
@@ -9,6 +9,21 @@ function twoDigits(d) {
   return d.toString();
 }
 
+function setTimeout_ (fn, delay) {
+    var maxDelay = Math.pow(2,31)-1;
+
+    if (delay > maxDelay) {
+        var args = arguments;
+        args[1] -= maxDelay;
+
+        return setTimeout(function () {
+            setTimeout_.apply(fn, args);
+        }, maxDelay);
+    }
+
+    return setTimeout.apply(fn, arguments);
+}
+
 module.exports = {
   name: "poll",
   description: "create a poll.",
@@ -94,7 +109,7 @@ module.exports = {
           error: ["time"]
         });
 
-        var options = optionString.first().content.split("\n");
+        var options = optionString.first().content.replace(/'/g, "#quot;").replace(/"/g, '#dquot;').split("\n");
         if (options.length <= 1) {
           return message.channel.send(
             "Please provide at least 2 options! Cancelling action..."
@@ -128,7 +143,7 @@ module.exports = {
             --num;
           }
         }
-        
+
         var currentDate = new Date();
         var newDate = new Date(currentDate.getTime() + time);
 
@@ -164,12 +179,15 @@ module.exports = {
           ":" +
           twoDigits(second) +
           " UTC";
-        
+        var pollMsg = "⬆**Poll**⬇";
         const Embed = new Discord.RichEmbed()
           .setColor(color)
           .setTitle(title)
           .setDescription(
-            "React with the numbers to vote!\nThis poll will end at:\n**" + readableTime + "**\n\n\n" + optionArray.join("\n\n")
+            "React with the numbers to vote!\nThis poll will end at:\n**" +
+              readableTime +
+              "**\n\n\n" +
+              optionArray.join("\n\n").replace(/#quot;/g, "'").replace(/#dquot;/g, '"')
           )
           .setTimestamp()
           .setFooter(
@@ -179,12 +197,10 @@ module.exports = {
               message.author.discriminator,
             message.author.displayAvatarURL
           );
-        var msg = await channel.send(Embed);
+        var msg = await channel.send(pollMsg, Embed);
         for (var i = 0; i < optionArray.length; i++) {
           await msg.react(emojis[i]);
         }
-
-        
 
         con.query(
           "INSERT INTO poll VALUES(" +
@@ -217,44 +233,55 @@ module.exports = {
           }
         );
 
-        setTimeout(async function() {
-          var result = [];
-          var end = [];
-          for (const emoji of msg.reactions.values()) {
-            await result.push(emoji.count);
-            var mesg =
-              "**" +
-              (emoji.count - 1) +
-              "** - `" +
-              allOptions[result.length - 1] +
-              "`";
-            await end.push(mesg);
-          }
-
-          const Ended = new Discord.RichEmbed()
-            .setColor(color)
-            .setTitle(title)
-            .setDescription(
-              "Poll ended. Here are the results:\n\n\n" + end.join("\n\n")
-            )
-            .setTimestamp()
-            .setFooter(
-              "Hosted by " +
-                message.author.username +
-                "#" +
-                message.author.discriminator,
-              message.author.displayAvatarURL
-            );
-          msg.edit(Ended);
-          msg.clearReactions().catch(err => {
-            console.error(err);
-          });
-          con.query("DELETE FROM poll WHERE id = " + msg.id, function(
+        setTimeout_(async function() {
+          con.query("SELECT * FROM poll WHERE id = " + msg.id, async function(
             err,
-            result
+            results,
+            fields
           ) {
             if (err) throw err;
-            console.log("Deleted an ended poll.");
+            if (results.length < 1) {
+              return;
+            } else {
+              var result = [];
+              var end = [];
+              for (const emoji of msg.reactions.values()) {
+                await result.push(emoji.count);
+                var mesg =
+                  "**" +
+                  (emoji.count - 1) +
+                  "** - `" +
+                  allOptions[result.length - 1] +
+                  "`";
+                await end.push(mesg);
+              }
+
+              const Ended = new Discord.RichEmbed()
+                .setColor(color)
+                .setTitle(title)
+                .setDescription(
+                  "Poll ended. Here are the results:\n\n\n" + end.join("\n\n").replace(/#quot;/g, "'").replace(/#dquot;/g, '"')
+                )
+                .setTimestamp()
+                .setFooter(
+                  "Hosted by " +
+                    message.author.username +
+                    "#" +
+                    message.author.discriminator,
+                  message.author.displayAvatarURL
+                );
+              msg.edit(pollMsg, Ended);
+              msg.clearReactions().catch(err => {
+                console.error(err);
+              });
+              con.query("DELETE FROM poll WHERE id = " + msg.id, function(
+                err,
+                result
+              ) {
+                if (err) throw err;
+                console.log("Deleted an ended poll.");
+              });
+            }
           });
         }, time);
 
@@ -300,19 +327,19 @@ module.exports = {
               "`";
             await end.push(mesg);
           }
-
+          var pollMsg = "⬆**Poll**⬇";
           const Ended = new Discord.RichEmbed()
             .setColor(color)
             .setTitle(result[0].title)
             .setDescription(
-              "Poll ended. Here are the results:\n\n\n" + end.join("\n\n")
+              "Poll ended. Here are the results:\n\n\n" + end.join("\n\n").replace(/#quot;/g, "'").replace(/#dquot;/g, '"')
             )
             .setTimestamp()
             .setFooter(
               "Hosted by " + author.username + "#" + author.discriminator,
               author.displayAvatarURL
             );
-          msg.edit(Ended);
+          msg.edit(pollMsg, Ended);
           msg.clearReactions().catch(err => {
             console.error(err);
           });
